+++ packages/keyring-api/src/api/v2/index.ts
@@ -4,4 +4,3 @@ export * from './keyring-type';
 export * from './create-account';
 export * from './export-account';
 export * from './private-key';
-export * from './wrapper';